# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AssetAdmission": {
        "properties": {
          "asset": {
            "$ref": "#/components/schemas/AssetView"
          },
          "ingestion": {
            "$ref": "#/components/schemas/IngestionView"
          }
        },
        "required": [
          "asset",
          "ingestion"
        ],
        "title": "AssetAdmission",
        "type": "object"
      },
      "AssetStatus": {
        "enum": [
          "quarantined",
          "queued",
          "processing",
          "ready",
          "partial",
          "failed",
          "cancelled"
        ],
        "title": "AssetStatus",
        "type": "string"
      },
      "AssetView": {
        "properties": {
          "byte_count": {
            "title": "Byte Count",
            "type": "integer"
          },
          "cloud_media_allowed": {
            "default": false,
            "title": "Cloud Media Allowed",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "mime_type": {
            "title": "Mime Type",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AssetStatus"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          }
        },
        "required": [
          "id",
          "name",
          "mime_type",
          "byte_count",
          "sha256",
          "status",
          "created_at"
        ],
        "title": "AssetView",
        "type": "object"
      },
...
      "Body_create_asset_api_v2_assets_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_create_asset_api_v2_assets_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IngestionStage": {
        "enum": [
          "quarantined",
          "queued",
          "parsing",
          "enriching",
          "indexing",
          "ready",
          "partial",
          "failed",
          "cancelled"
        ],
        "title": "IngestionStage",
        "type": "string"
      },
      "IngestionStatus": {
        "enum": [
          "queued",
          "running",
          "ready",
          "partial",
          "failed",
          "cancelled"
        ],
        "title": "IngestionStatus",
        "type": "string"
      },
      "IngestionView": {
        "properties": {
          "asset_id": {
            "format": "uuid",
            "title": "Asset Id",
            "type": "string"
          },
          "cancellation_requested": {
            "default": false,
            "title": "Cancellation Requested",
            "type": "boolean"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "document_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_seq": {
            "default": 0,
            "title": "Last Seq",
            "type": "integer"
          },
          "lexical_ready": {
            "default": false,
            "title": "Lexical Ready",
            "type": "boolean"
          },
          "semantic_ready": {
            "default": false,
            "title": "Semantic Ready",
            "type": "boolean"
          },
          "stage": {
            "$ref": "#/components/schemas/IngestionStage"
          },
          "status": {
            "$ref": "#/components/schemas/IngestionStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "id",
          "asset_id",
          "status",
          "stage",
          "created_at",
          "updated_at"
        ],
        "title": "IngestionView",
        "type": "object"
      },
      "MediaFrameView": {
        "properties": {
          "asset_id": {
            "format": "uuid",
            "title": "Asset Id",
            "type": "string"
          },
          "content_hash": {
            "title": "Content Hash",
            "type": "string"
          },
          "content_url": {
            "title": "Content Url",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "extraction_version_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Version Id"
          },
          "height": {
            "title": "Height",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "presentation_time_ms": {
            "title": "Presentation Time Ms",
            "type": "integer"
          },
          "rendition_id": {
            "format": "uuid",
            "title": "Rendition Id",
            "type": "string"
          },
          "source_kind": {
            "enum": [
              "periodic",
              "scene"
            ],
            "title": "Source Kind",
            "type": "string"
          },
          "width": {
            "title": "Width",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "asset_id",
          "rendition_id",
          "presentation_time_ms",
          "source_kind",
          "width",
          "height",
          "content_hash",
          "content_url",
          "created_at"
        ],
        "title": "MediaFrameView",
        "type": "object"
      },
      "MediaStoryboardView": {
        "properties": {
          "asset_id": {
            "format": "uuid",
            "title": "Asset Id",
            "type": "string"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "frames": {
            "items": {
              "$ref": "#/components/schemas/MediaFrameView"
            },
            "title": "Frames",
            "type": "array"
          },
          "sampled_times_ms": {
            "items": {
              "type": "integer"
            },
            "title": "Sampled Times Ms",
            "type": "array"
          },
          "sampling_strategy": {
            "title": "Sampling Strategy",
            "type": "string"
          },
          "tracks": {
            "items": {
              "$ref": "#/components/schemas/MediaTrackView"
            },
            "title": "Tracks",
            "type": "array"
          },
          "visual_coverage_warning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visual Coverage Warning"
          },
          "waveform_peaks": {
            "items": {
              "type": "number"
            },
            "title": "Waveform Peaks",
            "type": "array"
          }
        },
        "required": [
          "asset_id",
          "sampling_strategy"
        ],
        "title": "MediaStoryboardView",
        "type": "object"
      },
      "MediaTrackView": {
        "properties": {
          "asset_id": {
            "format": "uuid",
            "title": "Asset Id",
            "type": "string"
          },
          "average_frame_rate": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Frame Rate"
          },
          "channels": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channels"
          },
          "codec_name": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Codec Name"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "duration_ms": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "extraction_version_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Version Id"
          },
          "height": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "language": {
            "anyOf": [
              {
                "maxLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "sample_rate": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Rate"
          },
          "stream_index": {
            "minimum": 0.0,
            "title": "Stream Index",
            "type": "integer"
          },
          "track_type": {
            "enum": [
              "audio",
              "video"
            ],
            "title": "Track Type",
            "type": "string"
          },
          "width": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          }
        },
        "required": [
          "track_type",
          "stream_index",
          "id",
          "asset_id",
          "created_at"
        ],
        "title": "MediaTrackView",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ARES API",
    "version": "0.11.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/artifacts/{artifact_id}": {
      "get": {
        "operationId": "download_artifact_api_v1_artifacts__artifact_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Artifact"
      }
    },
...
    "/api/v2/assets": {
      "get": {
        "operationId": "list_assets_api_v2_assets_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AssetView"
                  },
                  "title": "Response List Assets Api V2 Assets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Assets"
      },
      "post": {
        "operationId": "create_asset_api_v2_assets_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_asset_api_v2_assets_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetAdmission"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Asset"
      }
    },
...
    "/api/v2/assets/{asset_id}/storyboard": {
      "get": {
        "operationId": "get_media_storyboard_api_v2_assets__asset_id__storyboard_get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaStoryboardView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Media Storyboard"
      }
    },
...
  }
}